requirements.txt
django-cors-headers
djangorestframework
Pillow